Migrated UpsConstants code to core/constants/ups
